+++ src/transform_lzo.c
@@ -153,11 +153,11 @@ static void bench_lzo_teardown (void *ctx) {
     free(ctx);
 }
 
-static uint64_t bench_lzo_comp_run (
+static size_t bench_lzo_comp_run (
     void *_ctx,
     const void *data_in,
-    const uint64_t data_in_size,
-    uint64_t *bytes_in
+    const size_t data_in_size,
+    size_t *bytes_in
 ) {
     struct bench_ctx *ctx = (struct bench_ctx *)_ctx;
